* 'pri_translation' of https://github.com/stoeckmann/util-linux:
  Avoid PRI specifiers in translated messages

+++ disk-utils/cfdisk.c
@@ -1778,8 +1778,8 @@ static int ui_refresh(struct cfdisk *cf)
 	attron(A_BOLD);
 	ui_center(0, _("Disk: %s"), fdisk_get_devname(cf->cxt));
 	attroff(A_BOLD);
-	ui_center(1, _("Size: %s, %"PRIu64" bytes, %ju sectors"),
-			strsz, bytes, (uintmax_t) fdisk_get_nsectors(cf->cxt));
+	ui_center(1, _("Size: %s, %ju bytes, %ju sectors"),
+			strsz, (uintmax_t) bytes, (uintmax_t) fdisk_get_nsectors(cf->cxt));
 	if (fdisk_get_disklabel_id(cf->cxt, &id) == 0 && id)
 		ui_center(2, _("Label: %s, identifier: %s"),
 				fdisk_label_get_name(lb), id);
@@ -1963,7 +1963,7 @@ static int ui_get_size(struct cfdisk *cf,	/* context */
 			if (insec)
 				user *= fdisk_get_sector_size(cf->cxt);
 			if (user < low) {
-				ui_warnx(_("Minimum size is %"PRIu64" bytes."), low);
+				ui_warnx(_("Minimum size is %ju bytes."), (uintmax_t) low);
 				rc = -ERANGE;
 			}
 			if (user > up && pwr && user < up + (1ULL << pwr * 10))
@@ -1972,7 +1972,7 @@ static int ui_get_size(struct cfdisk *cf,	/* context */
 				user = up;
 
 			if (user > up) {
-				ui_warnx(_("Maximum size is %"PRIu64" bytes."), up);
+				ui_warnx(_("Maximum size is %ju bytes."), (uintmax_t) up);
 				rc = -ERANGE;
 			}
 			if (rc == 0 && insec && expsize)
